Add support to changes in the jmri protocol and bump the version
The new version of ED 256f added changes to the trottle protocol passing
numbers instead of letters. Instead of MT+ -> M1+.
Added changes to support those.
Remove bug which was forcing the ED understand the loco was taken and should steal.
Bump the version to 1.7
Add script to create update
Add update file